src/main/java/com/example/demo/controller/PlayerController.java
package com.example.demo.controller;

import com.example.demo.persistance.Player;
import com.example.demo.service.player.PlayerService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
@RequestMapping("api/player")
public class PlayerController {

    private PlayerService playerService;

    @Autowired
    public PlayerController(PlayerService playerService) {
        this.playerService = playerService;
    }

    @GetMapping("/getAll")
    public List<Player> getAllPlayers(@RequestParam(value = "name", defaultValue = "World") String name) {
        return playerService.getAllPlayers();
    }
}
